Update bot files
Replace bot.js with bot-new.js and remove config.js
