--- config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  'ownerId'           : '<@!USERID>',
-  'botClientId'       : 'CLIENT_ID',
-  'botClientSecret'   : 'CLIENT_SECRET',
-  'botUsername'       : 'BOT_NAME',
-  'botToken'          : 'BOT_TOKEN',
-  'plexHostname'      : 'PLEX_IP',
-  'plexPort'          : 'PLEX_PORT',
-  'plexId'            : 'PLEX_ID',
-  'plexUsername'      : 'PLEX_USERNAME',
-  'plexPassword'      : 'PLEX_PASSWORD',
-  'plexToken'         : 'PLEX_TOKEN',
-  'omdbKey'           : 'OMDB_KEY',
-  'imgurClientId'     : 'IMGUR_CLIENT_ID',
-  'imgurClientSecret' : 'IMGUR_CLIENT_SECRET',
-  'imgurAlbum'        : 'IMGUR_ALBUM_URL',
-  'itadKey'           : 'ITAD_KEY',
-  'youtubeKey'        : 'YOUTUBE_KEY'
-};
